clean, add comments
