+++ consensus/state_sync/helpers.go
@@ -1,8 +1,6 @@
 package state_sync
 
 import (
-	"fmt"
-
 	typesCons "github.com/pokt-network/pocket/consensus/types"
 	cryptoPocket "github.com/pokt-network/pocket/shared/crypto"
 	"google.golang.org/protobuf/types/known/anypb"
@@ -10,8 +8,8 @@ import (
 
 // Helper function for broadcasting state sync messages to the all peers known to the node:
 //
-//		requests for metadata using the `periodicMetadataSynch()` function
-//	 	requests for blocks using the `StartSynching()` function
+//		sends metadata requests, via `metadataSyncLoop()` function
+//	 	sends block requests, via `()` function
 func (m *stateSync) broadcastStateSyncMessage(stateSyncMsg *typesCons.StateSyncMessage, block_height uint64) error {
 	m.logger.Info().Msg("📣 Broadcasting state sync message... 📣")
 
@@ -25,15 +23,10 @@ func (m *stateSync) broadcastStateSyncMessage(stateSyncMsg *typesCons.StateSyncM
 	// TODO: Use RainTree for this
 	// IMPROVE: OPtimize so this is not O(n^2)
 	for _, val := range validators {
-		//validatorAddr := val.GetAddress()
-		// DISCUSS_IN_THIS_COMMIT: You shouldn't need to do this check at the consensus module level - it's a P2P thin.
-		//	if m.GetBus().GetConsensusModule().GetNodeAddress() != validatorAddr {
 		if err := m.SendStateSyncMessage(stateSyncMsg, cryptoPocket.AddressFromString(val.GetAddress()), block_height); err != nil {
 			return err
 		}
-		//	}
 	}
-
 	return nil
 }
 
@@ -43,7 +36,6 @@ func (m *stateSync) SendStateSyncMessage(msg *typesCons.StateSyncMessage, dst cr
 	if err != nil {
 		return err
 	}
-	fmt.Println("Sending to: %s, I am %s", dst.ToString(), m.bus.GetConsensusModule().GetNodeAddress())
 	if err := m.GetBus().GetP2PModule().Send(dst, anyMsg); err != nil {
 		m.logger.Error().Err(err).Msg(typesCons.ErrSendMessage.Error())
 		return err
@@ -51,8 +43,7 @@ func (m *stateSync) SendStateSyncMessage(msg *typesCons.StateSyncMessage, dst cr
 	return nil
 }
 
-// TODO! remove this
-func (m *stateSync) logHelper(receiverPeerAddress string) map[string]any {
+func (m *stateSync) StateSyncLogHelper(receiverPeerAddress string) map[string]any {
 	consensusMod := m.GetBus().GetConsensusModule()
 
 	return map[string]any{
@@ -60,24 +51,4 @@ func (m *stateSync) logHelper(receiverPeerAddress string) map[string]any {
 		"senderPeerAddress":   consensusMod.GetNodeAddress(),
 		"receiverPeerAddress": receiverPeerAddress,
 	}
-
 }
-
-// func (m *stateSync) SendStateSyncMessage(stateSyncMsg *typesCons.StateSyncMessage, peerAddress cryptoPocket.Address, height uint64) error {
-// 	anyMsg, err := anypb.New(stateSyncMsg)
-// 	if err != nil {
-// 		return err
-// 	}
-
-// 	m.logger.Info().Fields(m.logHelper(peerAddress.ToString())).Msg("Sending StateSync Message")
-// 	return m.sendToPeer(anyMsg, peerAddress)
-// }
-
-// Helper function for messages to the peers
-// func (m *stateSync) sendToPeer(msg *anypb.Any, peerAddress cryptoPocket.Address) error {
-// 	if err := m.GetBus().GetP2PModule().Send(peerAddress, msg); err != nil {
-// 		m.logger.Error().Msgf(typesCons.ErrSendMessage.Error(), err)
-// 		return err
-// 	}
-// 	return nil
-// }
